refactor doctors search
