+++ frontend/src/components/AvailableDoctors.jsx
@@ -4,7 +4,7 @@ import axios from 'axios';
 import ClinicsList from './ClinicsList/ClinicsList';
 import SearchClinicsByAddressForm from './SearchClinicsByAddressForm';
 import MapComponent from './MapComponent';
-import { calculateCenter } from '../helpers/calcCenter';
+import SearchClinicsByDoctorForm from './SearchClinicsByDoctorForm';
 
 const defaultCenter = {
   lat: 43.642567, // default latitude
@@ -16,13 +16,11 @@ const AvailableDoctors = () => {
   const [clinics, setClinics] = useState([]);
   const [doctors, setDoctors] = useState([]);
   const [coordinates, setCoordinates] = useState(defaultCenter);
-  const [searchTerm, setSearchTerm] = useState('');
   const [displayedClinics, setDisplayedClinics] = useState([])
 
   const { isLoaded } = useJsApiLoader({
     googleMapsApiKey: process.env.REACT_APP_GOOGLE_MAPS_API_KEY,
   });
-
 
   useEffect(() => {
     if (isLoaded) {
@@ -44,38 +42,13 @@ const AvailableDoctors = () => {
     }
   }, [isLoaded]);
 
-  useEffect(() => {
-    if (searchTerm) {
-      const filteredClinics = clinics.filter(clinic => {
-        const doctor = doctors.find(doc => doc.clinic_id === clinic.id);
-        return doctor && doctor.name.toLowerCase().includes(searchTerm);
-      });
-      setDisplayedClinics(filteredClinics);
-      // Calculate the center of filtered clinics
-      const filteredClinicsCenter = calculateCenter(filteredClinics, defaultCenter);
-      setCoordinates(filteredClinicsCenter);
-    } else {
-      setDisplayedClinics(clinics);
-      setCoordinates(defaultCenter);
-    }
-  }, [searchTerm]);
-
-  
-
   if (!isLoaded) {
     return <div>Loading...</div>;
   }
 
   return (
     <div>
-      <form>
-        <input
-          type="text"
-          value={searchTerm}
-          onChange={(e) => setSearchTerm(e.target.value)}
-          placeholder="Enter doctor's name"
-        />
-      </form>
+      <SearchClinicsByDoctorForm clinics={clinics} doctors={doctors} setDisplayedClinics={setDisplayedClinics} setCoordinates={setCoordinates} defaultCenter={defaultCenter}/>
       <SearchClinicsByAddressForm setCoordinates={setCoordinates} setSearchTermMarker={setSearchTermMarker}/>
       <MapComponent clinics={displayedClinics} coordinates={coordinates} searchTermMarker={searchTermMarker}/>
       <ClinicsList clinics={displayedClinics} doctors={doctors} searchCoordinates={coordinates} />
